use SchemaValue instead of `any` for deepEqual values
