+++ next/src/types.ts
@@ -8,7 +8,7 @@ export type JsfSchemaType = Exclude<JSONSchema, boolean>['type']
 /**
  * Defines the type of a value in the form that will be validated against the schema.
  */
-export type SchemaValue = string | number | ObjectValue | null | undefined
+export type SchemaValue = string | number | ObjectValue | null | undefined | Array<SchemaValue>
 
 /**
  * A nested object value.
